+++ api/v1alpha1/jvmchaos_webhook_test.go
@@ -61,7 +61,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "",
 				},
@@ -84,7 +85,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateUpdate(chaos)
+						_, err := chaos.ValidateUpdate(chaos)
+						return err
 					},
 					expect: "",
 				},
@@ -107,7 +109,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateDelete()
+						_, err := chaos.ValidateDelete()
+						return err
 					},
 					expect: "",
 				},
@@ -129,7 +132,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},
@@ -151,7 +155,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},
@@ -173,7 +178,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},
@@ -195,7 +201,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},
@@ -217,7 +224,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},
@@ -233,7 +241,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},
@@ -249,7 +258,8 @@ var _ = Describe("jvmchaos_webhook", func() {
 						},
 					},
 					execute: func(chaos *JVMChaos) error {
-						return chaos.ValidateCreate()
+						_, err := chaos.ValidateCreate()
+						return err
 					},
 					expect: "error",
 				},