+++ CHANGELOG.md
@@ -23,6 +23,12 @@ For more information and how-to, see [RFC: Keep A Changelog](https://github.com/
 - Set the database column type to text for 'Workflow/Experiment/Schedule' [#4151](https://github.com/chaos-mesh/chaos-mesh/pull/4151)
 - Bump go to v1.20 [#4157](https://github.com/chaos-mesh/chaos-mesh/pull/4157)
 - Upgrade docker/login-action to v2 [#4167](https://github.com/chaos-mesh/chaos-mesh/pull/4167)
+- Update k8s dependencies to 1.28.0 [#4154](https://github.com/chaos-mesh/chaos-mesh/pull/4154)
+- Update lru dependency [#4189](https://github.com/chaos-mesh/chaos-mesh/pull/4189)
+- Update swag dependency [#4191](https://github.com/chaos-mesh/chaos-mesh/pull/4191)
+- Update ginkgo to 2.12.0 [#4190](https://github.com/chaos-mesh/chaos-mesh/pull/4190)
+- Update k8s controller-runtime dependency [#4198](https://github.com/chaos-mesh/chaos-mesh/pull/4198)
+- Automatically remove the token from the dashboard when it expires [#4193](https://github.com/chaos-mesh/chaos-mesh/pull/4193)
 
 ### Deprecated